Microservice/account-service-participant/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.21.0

